Rilis v2501.0.0

+++ donjo-app/controllers/Status_desa.php
@@ -43,7 +43,7 @@ class Status_desa extends Admin_Controller
 {
     public $modul_ini           = 'info-desa';
     public $sub_modul_ini       = 'status-desa';
-    public $kategori_pengaturan = 'status sdgs';
+    public $kategori_pengaturan = 'Status SDGs';
 
     public function __construct()
     {
@@ -80,21 +80,27 @@ class Status_desa extends Admin_Controller
 
             // Cek server Kemendes sebelum hapus cache
             try {
-                $client = new GuzzleHttp\Client();
-                $client->get(config_item('api_idm') . "/{$kode_desa}/{$tahun}", [
+                $client   = new GuzzleHttp\Client();
+                $response = $client->get(config_item('api_idm') . "/{$kode_desa}/{$tahun}", [
                     'headers' => [
                         'X-Requested-With' => 'XMLHttpRequest',
                     ],
                     'verify' => false,
                 ]);
+            } catch (Exception $e) {
+                log_message('error', $e->getMessage());
 
+                redirect_with('error', 'Tidak dapat mengambil data IDM.');
+            }
+
+            if ($response?->getStatusCode() === 200 && ($response->getBody()->getContents() !== '' && $response->getBody()->getContents() !== '0')) {
                 $this->cache->file->delete($cache);
                 set_session('tahun', $tahun);
 
                 redirect_with('success', 'Berhasil Perbarui Data');
-            } catch (Exception $e) {
-                log_message('error', $e->getMessage());
             }
+
+            redirect_with('error', 'Tidak dapat mengambil data IDM.');
         }
 
         redirect_with('error', 'Tidak dapat mengambil data IDM.');
